100 Ages -- Core Site Functionality
Core files for the "100 Ages, A Century of Voices" project. This
project published on July 4, 2013 and features a collection of
interviews with Boone County residents of all ages. It was published by
the Columbia Missourian and built using Angular.js. Font and image
files have been excluded from this commit. You can see the site here:
http://media.columbiamissourian.com/100ages/#/